Chapter five

+++ ch05_instructions/src/instructions/base/mod.rs
@@ -3,7 +3,7 @@ pub mod instruction;
 
 use super::super::rtda::Frame;
 
-pub fn branch(frame: &mut Frame, offset: i32) {
+pub fn branch(frame: &mut Frame, offset: i64) {
 	let pc = frame.get_thread().borrow().pc();
 	let next_pc = pc + offset;
 	frame.set_next_pc(next_pc);
